+++ docs/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.7](https://github.com/f3rnox/muzical/compare/v0.0.6...v0.0.7) (2026-04-23)
+
+
+### Features
+
+* add config, playlist, now_playing views ([605ce11](https://github.com/f3rnox/muzical/commit/605ce1112013ea95eb2f6653cd9969fa44867c4c))
+* add tests ([3f10331](https://github.com/f3rnox/muzical/commit/3f10331cadb73e91589ec74c899b55fb8ecfe267))
+* add volume controls ([0d3a566](https://github.com/f3rnox/muzical/commit/0d3a56646f9e269895876844a1ba4fc842d847f4))
+* update README ([16b2b18](https://github.com/f3rnox/muzical/commit/16b2b1812eab786c9f0300284b461fc51aefc1fc))
+
+
+### Bug Fixes
+
+* hide the playback bar in the now-playing view ([8dd920a](https://github.com/f3rnox/muzical/commit/8dd920ae2ec6ad91c6e728f5f8e2afdee906a303))
+
 ### [0.0.6](https://github.com/f3rnox/muzical/compare/v0.0.5...v0.0.6) (2026-04-22)